requirements.txt
# audioop-lts==0.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
gpt4all==2.8.2
idna==3.10
iniconfig==2.0.0
MouseInfo==0.1.3
packaging==24.2
pluggy==1.5.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.3.4
python-dotenv==1.0.1
python3-xlib==0.15
pyttsx3==2.98
pytweening==1.2.0
requests==2.32.3
setuptools==75.8.0
SpeechRecognition==3.14.1
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
vosk==0.3.45
websockets==14.2
